functions/pubsub_consumer/requirements.txt
google-cloud-pubsub>=2.10
google-cloud-bigquery>=3.0
python-dotenv>=0.21
